serverDjango/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
dj-database-url==2.0.0
Django==4.2.1
django-cors-headers==3.14.0
django-heroku==0.3.1
django-security==0.12.0
gunicorn==20.1.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
python-dateutil==2.8.0
python-dotenv==1.0.0
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.5.0
ua-parser==0.16.1
whitenoise==6.4.0
